still needs implementing of refresh token and cron job and needs to get tested again

+++ src/main/java/org/bookstore/bookstore/repositories/RefreshTokenRepository.java
@@ -27,7 +27,7 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
             value = "DELETE FROM RefreshTokens WHERE UserID = :userId",
             nativeQuery = true
     )
-    void deleteAllByUserId(@Param("userId") Long userId);
+    void deleteAllByUserId(@Param("userId") Integer userId);
